ATLAS/requirements.txt
# Core LangChain and LangGraph packages
langgraph>=0.2.0,<0.3.0
langchain>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0

# OpenAI API client
openai>=1.30.0,<2.0.0

# Environment management
python-dotenv>=1.0.0,<2.0.0

# Graph visualization
graphviz>=0.20.0,<1.0.0  # Pure Python package, easier to install
pygraphviz>=1.11,<2.0    # C extension, requires system graphviz